requirements.txt
fastapi
uvicorn
streamlit
chromadb
sentence-transformers
PyPDF2
python-multipart
pydantic
python-dotenv
requests
tqdm
groq-python
dataclasses-json
loguru
pytest
pytest-asyncio
httpx
pytest-mock